% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_analysis.R
\name{load_transcripts}
\alias{load_transcripts}
\title{Load Transcripts from a FASTA File}
\usage{
load_transcripts(file_path)
}
\arguments{
\item{file_path}{Path to the FASTA file.}
}
\value{
A DNAStringSet object containing the transcripts.
}
\description{
This function loads transcript sequences from a FASTA file using Biostrings.
}
